fuzz: address review feedback on dns/http fuzzers and unit tests
- fuzz_dns_parse / fuzz_http_parse now call the parsers directly
  (ndpi_search_dns / ndpi_search_http_tcp) with a synthesised
  packet_struct instead of driving ndpi_detection_process_packet,
  matching the fuzz_is_stun / fuzz_quic_get_crypto_data pattern.
- Expose ndpi_search_dns and ndpi_search_http_tcp via ndpi_private.h
  so fuzz/unit targets can reach them; wire -DNDPI_LIB_COMPILATION
  into the two fuzz target CFLAGS.
- Move hexDecodeUnitTest / tlsBlocksUnitTest (and the checkHexDecode,
  checkTlsBlocksDecode, rewriteHexCase helpers) from tests/unit/unit.c
  to example/utests.c; register them in run_unit_tests() so they run
  via ndpiReader --run-tests with the rest of the suite.
