+++ fuzz/fuzz_http_parse.c
@@ -2,18 +2,19 @@
  * fuzz_http_parse
  *
  * What it tests:
- *   HTTP dissection path reached from ndpi_detection_process_packet() for a
- *   TCP flow on port 80. Exercises ndpi_search_http_tcp and its header
+ *   HTTP dissector (ndpi_search_http_tcp) in src/lib/protocols/http.c. Calls
+ *   the parser directly with a synthesised packet_struct so the header
  *   extractors (Host, User-Agent, Content-Type, Referer), chunked/transfer
- *   handling, URL/URI walking, WebSocket upgrade, and the URL-decode helpers.
+ *   handling, URL/URI walking, WebSocket upgrade, and URL-decode helpers are
+ *   reached without relying on the full ndpi_detection_process_packet() flow
+ *   state machine.
  *
  * Expected input format:
- *   Raw TCP payload bytes (e.g. "GET / HTTP/1.1\r\nHost: x\r\n\r\n"). The
- *   harness wraps the fuzz data in a synthetic IPv4 + TCP packet with src/dst
- *   port 80 so the HTTP dissector fires via the port hint.
+ *   Raw TCP payload bytes (e.g. "GET / HTTP/1.1\r\nHost: x\r\n\r\n").
  */
 
 #include "ndpi_api.h"
+#include "ndpi_private.h"
 #include "fuzz_common_code.h"
 
 #include <arpa/inet.h>
@@ -21,65 +22,53 @@
 #include <stdio.h>
 #include <string.h>
 
-#define IPV4_HDR_LEN 20
-#define TCP_HDR_LEN  20
-#define MAX_PKT      (64 * 1024)
+static struct ndpi_detection_module_struct *ndpi_struct = NULL;
+static struct ndpi_flow_struct *ndpi_flow = NULL;
+static struct ndpi_iphdr iph;
+static struct ndpi_tcphdr tcph;
 
-static struct ndpi_detection_module_struct *ndpi_info_mod = NULL;
 static char *path = NULL;
 
 int LLVMFuzzerInitialize(int *argc, char ***argv) {
   (void)argc;
-  path = dirname(strdup(*argv[0]));
+  path = dirname(strdup(*argv[0])); /* No errors; no free! */
   return 0;
 }
 
 int LLVMFuzzerTestOneInput(const uint8_t *data, size_t size) {
-  static uint8_t pkt[MAX_PKT];
-  struct ndpi_flow_struct flow;
-  size_t payload_len, total_len;
-  uint16_t ip_len;
+  struct ndpi_packet_struct *packet;
 
-  if (ndpi_info_mod == NULL)
-    fuzz_init_detection_module(&ndpi_info_mod, NULL, path);
+  if (ndpi_struct == NULL) {
+    fuzz_init_detection_module(&ndpi_struct, NULL, path);
+    ndpi_flow = ndpi_calloc(1, sizeof(struct ndpi_flow_struct));
 
-  if (size > MAX_PKT - IPV4_HDR_LEN - TCP_HDR_LEN)
-    size = MAX_PKT - IPV4_HDR_LEN - TCP_HDR_LEN;
-  payload_len = size;
-  total_len = IPV4_HDR_LEN + TCP_HDR_LEN + payload_len;
+    memset(&iph, 0, sizeof(iph));
+    iph.version = 4;
+    iph.ihl = 5;
+    iph.protocol = 6; /* TCP */
+    iph.saddr = htonl(0x0A000001);
+    iph.daddr = htonl(0x0A000002);
 
-  memset(pkt, 0, IPV4_HDR_LEN + TCP_HDR_LEN);
-
-  /* IPv4 header: version=4, ihl=5, proto=6 (TCP) */
-  pkt[0] = 0x45;
-  ip_len = htons((uint16_t)total_len);
-  memcpy(&pkt[2], &ip_len, 2);
-  pkt[8] = 64;   /* ttl */
-  pkt[9] = 6;    /* TCP */
-  {
-    uint32_t saddr = htonl(0x0A000001);
-    uint32_t daddr = htonl(0x0A000002);
-    memcpy(&pkt[12], &saddr, 4);
-    memcpy(&pkt[16], &daddr, 4);
+    memset(&tcph, 0, sizeof(tcph));
+    tcph.source = htons(80);
+    tcph.dest = htons(80);
   }
 
-  /* TCP header: src/dst port 80; data offset = 5 (20 bytes) */
-  {
-    uint16_t sport = htons(80), dport = htons(80);
-    memcpy(&pkt[IPV4_HDR_LEN + 0], &sport, 2);
-    memcpy(&pkt[IPV4_HDR_LEN + 2], &dport, 2);
-  }
-  pkt[IPV4_HDR_LEN + 12] = 0x50;   /* data offset = 5 */
-  pkt[IPV4_HDR_LEN + 13] = 0x18;   /* flags = PSH | ACK */
+  fuzz_set_alloc_callbacks_and_seed(size);
 
-  if (payload_len > 0)
-    memcpy(&pkt[IPV4_HDR_LEN + TCP_HDR_LEN], data, payload_len);
+  packet = &ndpi_struct->packet;
+  packet->payload = data;
+  packet->payload_packet_len = (u_int16_t)size;
+  packet->iph = &iph;
+  packet->iphv6 = NULL;
+  packet->tcp = &tcph;
+  packet->udp = NULL;
 
-  memset(&flow, 0, SIZEOF_FLOW_STRUCT);
-  ndpi_detection_process_packet(ndpi_info_mod, &flow, pkt,
-                                (unsigned short)total_len, 0, NULL);
-  ndpi_detection_giveup(ndpi_info_mod, &flow);
-  ndpi_free_flow_data(&flow);
+  memset(ndpi_flow, 0, sizeof(struct ndpi_flow_struct));
+  ndpi_flow->l4_proto = IPPROTO_TCP;
+
+  ndpi_search_http_tcp(ndpi_struct, ndpi_flow);
+  ndpi_free_flow_data(ndpi_flow);
 
   return 0;
 }
